# url-shortener-backend/src/main/resources/api.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8080'
basePath: /
tags:
  - name: shorten-controller
    description: the shorten API
paths:
  /api/v1/shortening:
    get:
      tags:
        - List
      summary: List previous shortened URL
      description: List previous shortened URL
      operationId: listURL
      produces:
        - '*/*'
      parameters:
        - name: offset
          in: query
          required: false
          type: integer
          format: int64
        - name: pageNumber
          in: query
          required: false
          type: integer
          format: int32
        - name: pageSize
          in: query
          required: false
          type: integer
          format: int32
        - name: paged
          in: query
          required: false
          type: boolean
        - name: sort.sorted
          in: query
          required: false
          type: boolean
        - name: sort.unsorted
          in: query
          required: false
          type: boolean
        - name: unpaged
          in: query
          required: false
          type: boolean
      responses:
        '200':
          description: Expand successful
          schema:
            $ref: '#/definitions/ShortenUrlResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    post:
      tags:
        - Shorten
      summary: Input an url to shorten
      operationId: shortenUrl
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: url
          description: URL to be shortened
          required: true
          schema:
            $ref: '#/definitions/ShortenUrlRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ShortenUrlResponse'
        '201':
          description: Successful shortened URL
          schema:
            $ref: '#/definitions/ShortenUrlResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Invalid URL
      deprecated: false
  /api/v1/shortening/origin:
    get:
      tags:
        - Expand
      summary: Expand and redirect to origin URL
      description: Expand and redirect to origin URL
      operationId: expandURL
      produces:
        - '*/*'
      parameters:
        - name: shortenedUrl
          in: query
          description: Shorten URL
          required: true
          type: string
          allowEmptyValue: false
      responses:
        '200':
          description: OK
          schema:
            type: string
        '302':
          description: Expand successful
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
definitions:
  Pageable:
    type: object
    properties:
      offset:
        type: integer
        format: int64
      pageNumber:
        type: integer
        format: int32
      pageSize:
        type: integer
        format: int32
      paged:
        type: boolean
      sort:
        $ref: '#/definitions/Sort'
      unpaged:
        type: boolean
    title: Pageable
  Page«ShortenUrlResponse»:
    type: object
    properties:
      content:
        type: array
        items:
          $ref: '#/definitions/ShortenUrlResponse'
      empty:
        type: boolean
      first:
        type: boolean
      last:
        type: boolean
      number:
        type: integer
        format: int32
      numberOfElements:
        type: integer
        format: int32
      pageable:
        $ref: '#/definitions/Pageable'
      size:
        type: integer
        format: int32
      sort:
        $ref: '#/definitions/Sort'
      totalElements:
        type: integer
        format: int64
      totalPages:
        type: integer
        format: int32
    title: Page«ShortenUrlResponse»
  ShortenUrlRequest:
    type: object
    properties:
      url:
        type: string
    title: ShortenUrlRequest
  ShortenUrlResponse:
    type: object
    properties:
      createdAt:
        type: string
        format: date-time
      id:
        type: string
        format: uuid
      shortenedUrl:
        type: string
      url:
        type: string
    title: ShortenUrlResponse
  Sort:
    type: object
    properties:
      empty:
        type: boolean
      sorted:
        type: boolean
      unsorted:
        type: boolean
    title: Sort